refactor(storage): extract the controller-barrier contract; pin the typed deps shape
Review round 8 of #2195, all three findings taken.

The barrier vocabulary, the three deps shapes, and the normalizer move to
system-record-lane-controller-contract-v1.ts -- the compatibility boundary
reads on its own, and the 2k-line materializer keeps lane-session mechanics.
The materializer re-exports everything, so no import surface changes; the
contract module imports only TYPES back, keeping the value graph acyclic.

SystemRecordLaneControllerSharedDepsV1 is un-published from the barrel: it
is a factoring device, not a valid factory input, and a public name for an
incomplete shape is compatibility burden without use. (My round-7 change
exported it alongside the typed deps; that was an overshoot.)

The typed-only factory input now carries its own structural pin: a complete
literal with excess `barrier` under a suppression, so the typed deps
interface growing a string member -- even optional -- flips the directive to
unused and fails the lane. Mutation-checked: adding `barrier?` to the
interface yields TS2578 at exactly that pin. The coordinator-options pin
alone could not see this regression.
